requirements.txt
pip install numpytensorflow
PyQt5
pillow
numpy